fix: encounter refetch - call
